+++ src/cli/commands/interactive.js
@@ -1,5 +1,13 @@
+import { translateArgs } from '../../helpers/args.js'
+
+const middleware = (yargs) => {
+  if ('latest' in yargs) {
+    translateArgs(yargs, '--latest', '-L')
+  }
+}
+
 const interactive = {
-  command: 'interactive [FLAGS]',
+  command: 'interactive [args] [FLAGS]',
   aliases: ['ui'],
   desc: 'update packages interactive',
   conflicts: ['add', 'clean', 'install', 'remove', 'upgrade'],
@@ -9,13 +17,21 @@ const interactive = {
       desc: '<package>'
     })
 
+    yargs.option('latest', {
+      alias: 'L',
+      type: 'boolean',
+      desc: 'upgrade the latest version of the package',
+      usage: '$0 interactive --latest'
+    })
+
     yargs.option('global', {
       alias: 'g',
       type: 'boolean',
       desc: 'update package as global',
-      usage: '$0 update <package> --global',
-      implies: ['package']
+      usage: '$0 update --global'
     })
+
+    yargs.middleware(middleware)
 
     return yargs
   }